feat: Refactor common functions and http response code
