+++ authorizer.go
@@ -41,13 +41,11 @@ func NewAuthenticator() *Authorizer {
 
 func (a *Authorizer) TokenAuthMiddleware(h http.Handler) http.Handler {
 	fn := func(w http.ResponseWriter, r *http.Request) {
-		token := r.Header.Get("Authorization")
-		values := strings.Split(token, "Bearer ")
-		if token == "" || len(values) != 2 {
+		token, ok := getToken(r.Header)
+		if !ok {
 			respondWithError(w, http.StatusUnauthorized, "API token required")
 			return
 		}
-		token = values[1]
 		correct, userId, userRole := a.verifyIDToken(token)
 		if !correct {
 			respondWithError(w, http.StatusUnauthorized, "Invalid API Token")
@@ -63,13 +61,11 @@ func (a *Authorizer) TokenAuthMiddleware(h http.Handler) http.Handler {
 
 func (a *Authorizer) TokenAndBackAuthMiddleware(h http.Handler) http.Handler {
 	fn := func(w http.ResponseWriter, r *http.Request) {
-		token := r.Header.Get("Authorization")
-		values := strings.Split(token, "Bearer ")
-		if token == "" || len(values) != 2 {
+		token, ok := getToken(r.Header)
+		if !ok {
 			respondWithError(w, http.StatusUnauthorized, "API token required")
 			return
 		}
-		token = values[1]
 		hasBackendToken := token == os.Getenv("BACKEND_TOKEN")
 		if hasBackendToken {
 			h.ServeHTTP(w, r)
@@ -88,13 +84,11 @@ func (a *Authorizer) TokenAndBackAuthMiddleware(h http.Handler) http.Handler {
 
 func (a *Authorizer) BackAuthMiddleware(h http.Handler) http.Handler {
 	fn := func(w http.ResponseWriter, r *http.Request) {
-		token := r.Header.Get("Authorization")
-		values := strings.Split(token, "Bearer ")
-		if token == "" || len(values) != 2 {
+		token, ok := getToken(r.Header)
+		if !ok {
 			respondWithError(w, http.StatusUnauthorized, "API token required")
 			return
 		}
-		token = values[1]
 		hasBackendToken := token == os.Getenv("BACKEND_TOKEN")
 		if !hasBackendToken {
 			respondWithError(w, http.StatusUnauthorized, "Invalid API Token")
@@ -107,16 +101,14 @@ func (a *Authorizer) BackAuthMiddleware(h http.Handler) http.Handler {
 
 func (a *Authorizer) GinTokenAuthMiddleware() gin.HandlerFunc {
 	return func(c *gin.Context) {
-		token := c.Request.Header.Get("Authorization")
-		values := strings.Split(token, "Bearer ")
-		if token == "" || len(values) != 2 {
-			ginRespondWithError(c, 401, "API token required")
+		token, ok := getToken(c.Request.Header)
+		if !ok {
+			ginRespondWithError(c, http.StatusUnauthorized, "API token required")
 			return
 		}
-		token = values[1]
 		correct, userId, userRole := a.verifyIDToken(token)
 		if !correct {
-			ginRespondWithError(c, 401, "Invalid API token")
+			ginRespondWithError(c, http.StatusUnauthorized, "Invalid API token")
 			return
 		}
 		c.Set("userId", userId)
@@ -127,13 +119,11 @@ func (a *Authorizer) GinTokenAuthMiddleware() gin.HandlerFunc {
 
 func (a *Authorizer) GinTokenAndBackAuthMiddleware() gin.HandlerFunc {
 	return func(c *gin.Context) {
-		token := c.Request.Header.Get("Authorization")
-		values := strings.Split(token, "Bearer ")
-		if token == "" || len(values) != 2 {
-			ginRespondWithError(c, 401, "API token required")
+		token, ok := getToken(c.Request.Header)
+		if !ok {
+			ginRespondWithError(c, http.StatusUnauthorized, "API token required")
 			return
 		}
-		token = values[1]
 		hasBackendToken := token == os.Getenv("BACKEND_TOKEN")
 		if hasBackendToken {
 			c.Next()
@@ -142,7 +132,7 @@ func (a *Authorizer) GinTokenAndBackAuthMiddleware() gin.HandlerFunc {
 			c.Set("role", userRole)
 			c.Next()
 		} else {
-			ginRespondWithError(c, 401, "Invalid API token")
+			ginRespondWithError(c, http.StatusUnauthorized, "Invalid API token")
 			return
 		}
 	}
@@ -150,80 +140,41 @@ func (a *Authorizer) GinTokenAndBackAuthMiddleware() gin.HandlerFunc {
 
 func (a *Authorizer) GinBackAuthMiddleware() gin.HandlerFunc {
 	return func(c *gin.Context) {
-		token := c.Request.Header.Get("Authorization")
-		values := strings.Split(token, "Bearer ")
-		if token == "" || len(values) != 2 {
-			ginRespondWithError(c, 401, "API token required")
+		token, ok := getToken(c.Request.Header)
+		if !ok {
+			ginRespondWithError(c, http.StatusUnauthorized, "API token required")
 			return
 		}
-		token = values[1]
 		hasBackendToken := token == os.Getenv("BACKEND_TOKEN")
 		if hasBackendToken {
 			c.Next()
 		} else {
-			ginRespondWithError(c, 401, "Invalid API token")
+			ginRespondWithError(c, http.StatusUnauthorized, "Invalid API token")
 			return
 		}
 	}
 }
 
-func (a *Authorizer) downloadCertificates() {
-	var certs map[string]string
-	res, err := http.Get(CertsAPIEndpoint)
-	if err != nil {
-		return
-	}
-	defer res.Body.Close()
-	data, err := ioutil.ReadAll(res.Body)
-	if err != nil {
-		return
-	}
-	json.Unmarshal(data, &certs)
-	a.Certs = certs
-}
+func (a *Authorizer) verifyJWT(t string) (claims jwt.MapClaims, ok bool) {
+	parsed, _ := jwt.Parse(t, func(t *jwt.Token) (interface{}, error) {
+		cert, _ := a.GetCertificateFromToken(t)
+		publicKey, err := readPublicKey(cert)
+		if err != nil {
+			return "", err
+		}
+		return publicKey, nil
+	})
 
-func ginRespondWithError(c *gin.Context, code int, message interface{}) {
-	c.AbortWithStatusJSON(code, gin.H{"error": message})
-}
-
-func respondWithError(w http.ResponseWriter, code int, message interface{}) {
-	w.Header().Set("Content-Type", "application/json")
-	w.WriteHeader(code)
-	err := make(map[string]interface{})
-	err["error"] = message
-	json.NewEncoder(w).Encode(err)
-}
-
-var CertsAPIEndpoint = "https://www.googleapis.com/robot/v1/metadata/x509/securetoken@system.gserviceaccount.com"
-
-func (a *Authorizer) getCertificates() (certs map[string]string) {
-	return a.Certs
-}
-
-func (a *Authorizer) getCertificate(kid string) (cert []byte) {
-	certs := a.getCertificates()
-	certString, ok := certs[kid]
+	ok = parsed.Valid
 	if !ok {
-		a.downloadCertificates()
-		certs := a.getCertificates()
-		certString = certs[kid]
+		return
 	}
-	cert = []byte(certString)
+	if parsed.Header["alg"] != "RS256" {
+		ok = false
+		return
+	}
+	claims, ok = verifyPayload(parsed)
 	return
-}
-
-func (a *Authorizer) GetCertificateFromToken(token *jwt.Token) ([]byte, error) {
-
-	// Get kid
-	kid, ok := token.Header["kid"]
-	if !ok {
-		return []byte{}, errors.New("kid not found")
-	}
-	kidString, ok := kid.(string)
-	if !ok {
-		return []byte{}, errors.New("kid cast error to string")
-	}
-	return a.getCertificate(kidString), nil
 }
 
 func (a *Authorizer) verifyIDToken(token string) (bool, string, string) {
@@ -294,24 +245,23 @@ func readPublicKey(cert []byte) (*rsa.PublicKey, error) {
 	return publicKey, nil
 }
 
-func (a *Authorizer) verifyJWT(t string) (claims jwt.MapClaims, ok bool) {
-	parsed, _ := jwt.Parse(t, func(t *jwt.Token) (interface{}, error) {
-		cert, _ := a.GetCertificateFromToken(t)
-		publicKey, err := readPublicKey(cert)
-		if err != nil {
-			return "", err
-		}
-		return publicKey, nil
-	})
+func getToken(headers http.Header) (string, bool) {
+	token := headers.Get("Authorization")
+	values := strings.Split(token, "Bearer ")
+	if token == "" || len(values) != 2 {
+		return "", false
+	}
+	return values[1], true
+}
 
-	ok = parsed.Valid
-	if !ok {
-		return
-	}
-	if parsed.Header["alg"] != "RS256" {
-		ok = false
-		return
-	}
-	claims, ok = verifyPayload(parsed)
-	return
+func ginRespondWithError(c *gin.Context, code int, message interface{}) {
+	c.AbortWithStatusJSON(code, gin.H{"error": message})
+}
+
+func respondWithError(w http.ResponseWriter, code int, message interface{}) {
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(code)
+	err := make(map[string]interface{})
+	err["error"] = message
+	json.NewEncoder(w).Encode(err)
 }
